chore: refactor and move badge helper functions to utils
